+++ ceammc/ext/src/base/is_file.h
@@ -1,5 +1,5 @@
 /*****************************************************************************
- * Copyright 2018 Serge Poltavsky. All rights reserved.
+ * Copyright 2020 Serge Poltavsky. All rights reserved.
  *
  * This file may be distributed under the terms of GNU Public License version
  * 3 (GPL v3) as defined by the Free Software Foundation (FSF). A copy of the
@@ -11,28 +11,24 @@
  * contact the author of this file, or the owner of the project in which
  * this file belongs to.
  *****************************************************************************/
-#ifndef IS_DATA_H
-#define IS_DATA_H
+#ifndef IS_FILE_H
+#define IS_FILE_H
 
+#include "ceammc_data.h"
 #include "ceammc_object.h"
 
 using namespace ceammc;
 
-class IsData : public BaseObject {
-    FlagProperty* in_list_;
-
+class IsFile : public BaseObject {
 public:
-    IsData(const PdArgs& args);
-    void onBang() override;
-    void onFloat(t_float f) override;
-    void onSymbol(t_symbol* s) override;
-    void onList(const AtomList& lst) override;
-    void onAny(t_symbol* s, const AtomList& lst) override;
-    void onData(const Atom& d) override;
+    IsFile(const PdArgs& a);
+    void onSymbol(t_symbol* s) final;
+    void onDataT(const StringAtom& str);
 
-    bool processAnyProps(t_symbol* sel, const AtomList& lst) final;
+private:
+    bool fileExists(const std::string& p) const;
 };
 
-void setup_is_data();
+void setup_is_file();
 
-#endif // IS_DATA_H
+#endif // IS_FILE_H